requirements.txt
bleak==3.0.1
pillow==12.2.0
psutil==7.2.2
pyserial==3.5
rich==15.0.0
typer==0.25.0
winsdk==1.0.0b10
